XJDFLIB-1: XJdfNodeFactory Comment XJDFLIB-2: Float to Double XJDFLIB-3: ContactBuilder
git-svn-id: http://svn.cip4.org/svnroot/trunk@5556 25ee8a6d-831b-0410-907a-ff8528d65d49

+++ src/main/java/org/cip4/lib/xjdf/util/ConversionUtil.java
@@ -83,11 +83,11 @@ public class ConversionUtil {
 
 	/**
 	 * Convert Millimeter to DTP-Points.
-	 * @param millimeter Millimeter as float value.
+	 * @param millimeter Millimeter as double value.
 	 * @return Millimeter value in DTP-Points.
 	 */
-	public static float mm2Dtp(float millimeter) {
-		return millimeter / 25.4f * 72f;
+	public static double mm2Dtp(double millimeter) {
+		return millimeter / 25.4d * 72d;
 	}
 
 	/**
@@ -95,8 +95,8 @@ public class ConversionUtil {
 	 * @param dtp DTP-Points as value.
 	 * @return DTP-Points value in Millimeter.
 	 */
-	public static float dtp2Mm(float dtp) {
-		return dtp / 72f * 25.4f;
+	public static double dtp2Mm(double dtp) {
+		return dtp / 72d * 25.4d;
 	}
 
 	/**
@@ -104,10 +104,10 @@ public class ConversionUtil {
 	 * @param dtp DTP-Points as String.
 	 * @return DTP-Points value in Millimeter.
 	 */
-	public static float dtp2MmS(String strDtp) {
+	public static double dtp2MmS(String strDtp) {
 
-		// convert string to float
-		float val = Float.valueOf(strDtp);
+		// convert string to double
+		double val = Double.valueOf(strDtp);
 
 		// compute
 		return dtp2Mm(val);
@@ -115,15 +115,15 @@ public class ConversionUtil {
 
 	/**
 	 * Convert Millimeter to DTP-Points String.
-	 * @param millimeter Millimeter as float value.
+	 * @param millimeter Millimeter as double value.
 	 * @return Millimeter value in DTP-Points as String.
 	 */
-	public static String mm2DtpS(float millimeter) {
+	public static String mm2DtpS(double millimeter) {
 
 		// convert
-		float dtp = mm2Dtp(millimeter);
+		double dtp = mm2Dtp(millimeter);
 
 		// result as string
-		return Float.toString(dtp);
+		return Double.toString(dtp);
 	}
 }
